REF: Rename `mutations` module to `api`
